Added null checks

+++ CookIt.API/Controllers/ImageController.cs
@@ -27,12 +27,20 @@ namespace CookIt.API.Controllers
         [HttpPost("getImage")]
         public async Task<IActionResult> GetScaledImageAsync(ImageResizeDto imageResize)
         {
+            if(imageResize.Src == null)
+            {
+                return BadRequest();
+            }
             string base64 = await this._imageService.GetOrSetScaledImageAsync(imageResize.Src, imageResize.Width, imageResize.Height);
             return Ok(base64);
         }
         [HttpDelete("deleteImage")]
         public IActionResult DeleteImage(ImageResizeDto imageResize)
         {
+            if (imageResize.Src == null )
+            {
+                return BadRequest();
+            }
             bool imageDeleted = this._imageService.DeleteImage(imageResize.Src, imageResize.Width, imageResize.Height);
             if (imageDeleted == false)
             {
